♻️ refactor(webapp): extract employee settings sections
