+++ apps/webapp/src/app/[locale]/(app)/settings/employees/[employeeId]/employee-section-shared.ts
@@ -1,0 +1,7 @@
+export type Translate = (
+	key: string,
+	defaultValue: string,
+	values?: Record<string, string | number>,
+) => string;
+
+export const defaultTranslate: Translate = (_key, defaultValue) => defaultValue;
